Configuração do database

+++ src/dto/user_dto.go
@@ -1,6 +1,6 @@
 package dto
 
 type User struct {
-	Id       int16  `json:"id"`
+	Id       int8   `json:"id"`
 	Username string `json:"username"`
 }
